requirements.txt
Flask==1.1.2
feedparser==6.0.2
beautifulsoup4==4.9.3
google-api-python-client==1.12.8
google-cloud-datastore==2.1.0
grpcio==1.29.0
python-dateutil==2.8.1
